+++ docs/releases.md
@@ -50,6 +50,7 @@ changes over time.
 * github-385: Update Flowman Tutorial
 * github-412: Create tutorial for using Flowman Maven Plugin
 * github-371: Automate EMR integration test via AWS CLI
+* github-431: Update EMR build profil to 6.12
 
 
 ### Version 1.0.1 - Upcoming